requirements.txt
django==5.0
djangorestframework==3.14.0
google-generativeai==0.3.0
python-dotenv==1.0.0
django-cors-headers==4.3.0
